+++ CHANGES.rst
@@ -8,6 +8,10 @@ Changes to be released in next version
  * Settings: The notifications toggle no longer detects the system's "Deliver Quietly" configuration as disabled (#2368).
  * Settings: Adds a link to open the Settings app to quickly configure app notifications.
  * VoIP: Text & icon changes on call tiles (#4642).
+ * Voice messages: Stop recording and go into locked mode when the application becomes inactive (#4656)
+ * Voice messages: Allow voice message playback control from the iOS lock screen and control center (#4655)
+ * Voice messages: Improve audio recording quality
+ * Voice messages: Remove labs setting and enable them by default
 
 🐛 Bugfix
  *